feat(matcher): E1.1 — scaffold exact_matching package + move therapy_match_profile
First slice of extracting the matcher into a self-contained, CB-importable
Django library (E1, epic cancerbot#4601). Scaffolds matcher_app/ (pyproject,
exact_matching package) and moves the pure trial-side therapy column-name profile
into exact_matching.therapy_match_profile. EXACT's trials/services/
therapy_match_profile.py becomes a back-compat shim re-exporting from the package,
so all importers (queryset, matcher, trial_attributes, promop_adapter,
therapy_graph, tests) stay behavior-identical. Package installed editable via
requirements (-e ./matcher_app).

QuerySet, per-trial matcher, config, abstract cluster models, and ExactMatcher
(CB's MatcherBackend impl) follow in E1.2-E1.4.

Full EXACT suite green (1073 passed, 5 skipped).

Part of EPIC cancerbot#4601 (E1 #272). Proposal — cross-team.
