+++ .gitignore
@@ -58,3 +58,9 @@ trials/management/commands/evaluate_ethalon_live.py
 tests/management/test_evaluate_ethalon_live.py
 scripts/evaluator/evaluate_live.sh
 republish_latest.sh
+
+# exact_matching package build artifacts
+matcher_app/*.egg-info/
+matcher_app/**/*.egg-info/
+matcher_app/build/
+matcher_app/dist/
