+++ matcher_app/exact_matching/__init__.py
@@ -1,0 +1,8 @@
+"""exact_matching — shared CB/EXACT trial matcher (extraction in progress, E1).
+
+E1.1 moves the pure trial-side therapy column-name profile
+(`therapy_match_profile`). The QuerySet, per-trial matcher, config, abstract
+models, and `ExactMatcher` (CB's MatcherBackend impl) follow in E1.2-E1.4.
+EXACT's in-tree modules re-export from here so ports stay behavior-identical.
+"""
+__version__ = "0.0.1"
