+++ requirements.txt
@@ -40,3 +40,4 @@ pytest==8.4.2
 pytest-django==4.11.1
 pytest-factoryboy==2.8.1
 pytest-split==0.10.0
+-e ./matcher_app
